split constants into separate files
