+++ src/constants/ids.ts
@@ -1,5 +1,3 @@
 export const emptyId = "00000000-0000-0000-0000-000000000000";
-export const spotlightInUseErrorCode = 20400;
-export const snippetInUseErrorCode = 204004;
 export const defaultName = "Default";
 export const defaultCodename = "default";
