+++ src/constants/responseCodes.ts
@@ -1,0 +1,2 @@
+export const spotlightInUseErrorCode = 20400;
+export const snippetInUseErrorCode = 204004;
